fix: fix deployments scripts
1) result of cd command is not checked so if there is no
   ../site-assets directory, we are puring the whole repo
   in the next command.
2) the path to the checked-out repo is named 'www' not 'code'

+++ deploy.sh
@@ -6,12 +6,16 @@ make
 cd ../..
 hugo
 cd ../site-assets
-git pull
-rm -rf ./*
-cp -r ../code/public/* .
-cp -r ../code/CNAME .
-git add .
-git commit -m "new deployment"
-git push
-cd ../code
-git pull
+if [ $? -ne 0 ]; then
+	echo "You don't have the ../site-assets directory"
+else
+	git pull
+	rm -rf ./*
+	cp -r ../www/public/* .
+	cp -r ../www/CNAME .
+	git add .
+	git commit -m "new deployment"
+	git push
+	cd ../www
+	git pull
+fi
